Fix scripts not loading in time on slower connections
The scripts were not loading in time on a slower connection than
gigabit, causing the browser to trry and execute the scripts before
actually recieving the scripts. There was no visible error in the
console causing some consfusion on why the scripts were not executing.

This commit fixes the matter by removing the async loader and executing
the scripts async using the async modifier in HTML.

To test this commit, load the site on your testing environment and throttle
the connection. this can be done in the Chrome Dev Tools. if the scripts are
loading correctly after a reload, this fix has worked.

Signed-off-by: MegaXLR <admin@megaxlr.net>